requirements/_base.txt
beautifulsoup4
django
django-tastypie
django-twitter
gunicorn
honcho
lxml
psycopg2
recaptcha-client
simplejson
